navbar with parent specification works :) reuse in settings-side
