+++ src/pages/navbar.html
@@ -2,8 +2,8 @@
   <!-- <template repeat.for="route of navModel.routes">
     <a href.bind="route.path | firstNonEmpty">${route.title}</a>
   </template> -->
-  <div repeat.for="route of navModel.routes">
-    <button fill="clear" load.bind="route.path | firstNonEmpty">${route.title}</button>
+  <div repeat.for="route of routes">
+    <button fill="clear" load.bind="getLink(route)">${route.title}</button>
     <br>
   </div>
 </nav>
