+++ src/pages/settings-page/settings-side.html
@@ -1,18 +1,6 @@
-
+<import from="../navbar"></import>
 <hr>
 hi settings side
 
-<nav style="display: flex; gap: 0.5rem;">
-  <!-- <template repeat.for="route of navModel.routes">
-    <a href.bind="route.path | firstNonEmpty">${route.title}</a>
-  </template> -->
+<navbar parent="settings"></navbar>
 
-  <!-- FIXME: use path instead of id. But need to be inside main's viewport for the subroute to apply this way, 
-   otherwise we get "url/account" instead of "url/settings/account" -->
-   <!-- route.id -->
-   <!-- route.data.parent + '/' + ((route.path | firstNonEmpty))) -->
-  <div repeat.for="route of routes">
-    <button fill="clear" load.bind="getLink(route)">${route.title}</button>
-    <br>
-  </div>
-</nav>
